improve footer templates
