+++ apps/meteor/client/sidebar/footer/voip/VoipFooter.tsx
@@ -2,7 +2,7 @@ import type { IVoipRoom } from '@rocket.chat/core-typings';
 import { ICallerInfo, VoIpCallerInfo, VoipClientEvents } from '@rocket.chat/core-typings';
 import { css } from '@rocket.chat/css-in-js';
 import { Box, Button, ButtonGroup, Icon, SidebarFooter, Menu, IconButton } from '@rocket.chat/fuselage';
-import React, { ReactElement, MouseEvent } from 'react';
+import React, { ReactElement, MouseEvent, ReactNode } from 'react';
 
 import { useVoipFooterMenu } from '../../../../ee/client/hooks/useVoipFooterMenu';
 import { CallActionsType } from '../../../contexts/CallContext';
@@ -32,6 +32,7 @@ type VoipFooterPropsType = {
 	openedRoomInfo: { v: { token?: string | undefined }; rid: string };
 	anonymousText: string;
 	isEnterprise: boolean;
+	children?: ReactNode;
 };
 
 export const VoipFooter = ({
@@ -52,6 +53,7 @@ export const VoipFooter = ({
 	openedRoomInfo,
 	anonymousText,
 	isEnterprise = false,
+	children,
 }: VoipFooterPropsType): ReactElement => {
 	const cssClickable =
 		callerState === 'IN_CALL' || callerState === 'ON_HOLD'
@@ -164,6 +166,7 @@ export const VoipFooter = ({
 					</ButtonGroup>
 				</Box>
 			</Box>
+			{children}
 		</SidebarFooter>
 	);
 };
