+++ apps/meteor/client/views/hooks/useUpgradeTabParams.ts
@@ -10,7 +10,7 @@ export const useUpgradeTabParams = (): { tabType: UpgradeTabVariant | false; tri
 	const cloudWorkspaceHadTrial = useSetting('Cloud_Workspace_Had_Trial') as boolean;
 
 	const { data: registrationStatusData } = useQuery(['registrationStatus'], () => getRegistrationStatus());
-	const { data: getValidLicensesData, isSuccess } = useQuery(['licences'], () => getLicenses(), {
+	const { data: getValidLicensesData, isSuccess } = useQuery(['licenses'], () => getLicenses(), {
 		enabled: !!registrationStatusData,
 	});
